update update update update update my precious

+++ kernel/include/x86_64/tss.h
@@ -7,7 +7,7 @@ typedef struct{
 	uint64_t rsp0;
 	uint64_t rsp1;
 	uint64_t rsp2;
-	uint32_t rsvd1;
+	uint64_t rsvd1;
 	uint64_t ist1;
 	uint64_t ist2;
 	uint64_t ist3;
